(SIMP-1736) Hiera-based parameter enforcement
This implements a Hiera backend that has the ability to stack various
compliance profiles in such a way that a unified parameter stack can be
applied successfully across *multiple* compliance data sets.

Both a Hiera v3 and a v5 backend are provided, and most of the
compliance code is shared between them.

This should be pulled in and used in acceptance testing across *ALL*
SIMP modules since it provides for rapid implementation testing of the
various components and expansion of the back-end data sets.

SIMP-1736 #close